Move LineageFilter to separate component.
